refactor: folder naming preview pdf for custom query with one folder
